+++ inc/lilyParser/index.ts
@@ -3,14 +3,18 @@
 //import lilyGrammar from "./lilypond.jison";
 //console.log("lilyGrammar:", lilyGrammar);
 
-const jison = require("jison");
+import jison from "jison";
 
 
 
 const createParser = grammar => new jison.Parser(grammar);
 
 
+const hookJisonPrint = (print = () => {}) => jison.print = print;
+
+
 
 export {
 	createParser,
+	hookJisonPrint,
 };