+++ inc/lilyParser/lilyDocument.ts
@@ -1,0 +1,16 @@
+
+const parseRaw = data => {
+	// TODO:
+	return data;
+};
+
+
+
+export default class LilyDocument {
+	root: any;
+
+
+	constructor (data) {
+		this.root = parseRaw(data);
+	}
+};